\name{marmap}
\alias{marmap}
\alias{marmap}
\docType{package}
\title{
Import, plot and analyze bathymetric and topographic data
}
\description{
Import, plot and analyze bathymetric and topographic data
}
\details{
\tabular{ll}{
Package: \tab marmap\cr
Type: \tab Package\cr
Version: \tab 0.8\cr
Date: \tab 2013-10-19\cr
}
Import, plot and analyze bathymetric and topographic data
}
\author{
Eric Pante and Benoit Simon-Bouhet

Maintainer: Eric Pante <pante.eric@gmail.com>
}
