\name{marmap}
\alias{marmap}
\alias{marmap}
\docType{package}
\title{
Import, plot and analyze bathymetric and topographic data
}
\description{
marmap is a package designed for downloading, plotting and manipulating bathymetric and topographic data in R. It can query the ETOPO1 bathymetry and topography database hosted by the NOAA, use simple latitude-longitude-depth data in ascii format, and take advantage of the advanced plotting tools available in R to build publication-quality bathymetric maps. Functions to query data (bathymetry, sampling information, etc...) are available interactively by clicking on marmap maps. Bathymetric and topographic data can also be used to calculate projected surface areas within specified depth/altitude intervals, and constrain the calculation of realistic shortest path distances.
}
\details{
\tabular{ll}{
Package: \tab marmap\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2018-09-22\cr
}
Import, plot and analyze bathymetric and topographic data
}

\references{
Pante E, Simon-Bouhet B (2013) marmap: A Package for Importing, Plotting and Analyzing Bathymetric and Topographic Data in R. PLoS ONE 8(9): e73051. doi:10.1371/journal.pone.0073051
}
\author{
Eric Pante, Benoit Simon-Bouhet and Jean-Olivier Irisson

Maintainer: Benoit Simon-Bouhet <besibo@gmail.com>
}
